+++ forge/src/main/java/earth/terrarium/adastra/client/forge/AdAstraClientForge.java
@@ -8,10 +8,12 @@ import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
 import net.minecraft.world.item.Item;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.*;
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -20,59 +22,62 @@ import java.util.Map;
 public class AdAstraClientForge {
     public static final Map<Item, BlockEntityWithoutLevelRenderer> ITEM_RENDERERS = new HashMap<>();
 
+    public static void init() {
+        FMLJavaModLoadingContext.get().getModEventBus().addListener(AdAstraClientForge::onSetupItemColors);
+        MinecraftForge.EVENT_BUS.addListener(AdAstraClientForge::onRegisterClientHud);
+        MinecraftForge.EVENT_BUS.addListener(AdAstraClientForge::onClientTick);
+        MinecraftForge.EVENT_BUS.addListener(AdAstraClientForge::onRenderLevelStage);
+    }
+
     @SubscribeEvent
     public static void onClientSetup(FMLClientSetupEvent event) {
         event.enqueueWork(AdAstraClient::init);
-        AdAstraClient.onRegisterItemRenderers(ITEM_RENDERERS::put);
+        AdAstraClient.onRegisterItemRenderers(ITEM_RENDERERS::put); //
     }
 
     @SubscribeEvent
     public static void onRegisterKeyBindings(RegisterKeyMappingsEvent event) {
-        event.register(AdAstraClient.KEY_TOGGLE_SUIT_FLIGHT);
+        event.register(AdAstraClient.KEY_TOGGLE_SUIT_FLIGHT); //
         event.register(AdAstraClient.KEY_OPEN_RADIO);
     }
 
     @SubscribeEvent
     public static void onRegisterParticles(RegisterParticleProvidersEvent event) {
-        AdAstraClient.onRegisterParticles((type, provider) -> event.registerSpriteSet(type, provider::create));
+        AdAstraClient.onRegisterParticles((type, provider) -> event.registerSpriteSet(type, provider::create)); //
     }
 
     @SubscribeEvent
     public static void modelLoading(ModelEvent.RegisterAdditional event) {
-        AdAstraClient.onRegisterModels(event::register);
+        AdAstraClient.onRegisterModels(event::register); //
     }
 
     @SubscribeEvent
     public static void onRegisterLayerDefinitions(EntityRenderersEvent.RegisterLayerDefinitions event) {
-        AdAstraClient.onRegisterEntityLayers(event::registerLayerDefinition);
+        AdAstraClient.onRegisterEntityLayers(event::registerLayerDefinition); //
     }
 
     @SubscribeEvent
     public static void onRegisterDimensionSpecialEffects(RegisterDimensionSpecialEffectsEvent event) {
-        ModDimensionRenderers.onRegisterDimensionSpecialEffects((dimension, effects) -> event.register(dimension.location(), effects));
+        ModDimensionRenderers.onRegisterDimensionSpecialEffects((dimension, effects) -> event.register(dimension.location(), effects)); //
     }
 
-    @SubscribeEvent
-    public void onClientTick(TickEvent.ClientTickEvent event) {
+    private static void onClientTick(TickEvent.ClientTickEvent event) {
         if (event.phase.equals(TickEvent.Phase.START)) {
             AdAstraClient.clientTick(Minecraft.getInstance());
         }
     }
 
-    @SubscribeEvent
-    public void onRenderLevelStage(RenderLevelStageEvent event) {
+    private static void onRenderLevelStage(RenderLevelStageEvent event) {
         if (event.getStage() == RenderLevelStageEvent.Stage.AFTER_PARTICLES) {
             OxygenDistributorOverlayRenderer.render(event.getPoseStack(), event.getCamera());
         }
     }
 
-    @SubscribeEvent
-    public void onRegisterClientHud(RenderGuiEvent.Post event) {
+    private static void onRegisterClientHud(RenderGuiEvent.Post event) {
         AdAstraClient.onRegisterHud(hud -> hud.renderHud(event.getGuiGraphics(), event.getPartialTick()));
     }
 
-    @SubscribeEvent
-    public void onSetupItemColors(RegisterColorHandlersEvent.Item event) {
+    private static void onSetupItemColors(RegisterColorHandlersEvent.Item event) {
         AdAstraClient.onAddItemColors(event::register);
     }
 }